[INSCON-13] removed use of null coalescing operator and fixed unit test cases that were affected by the change. Return types should now be consistent for cached and non-cached objects.

+++ Enterprise/DAL/Enterprise.DAL.Core/Service/LookupService.cs
@@ -37,7 +37,7 @@ namespace Enterprise.DAL.Core.Service
             var record = new Lookup
                 {
                     LookupID = reader.GetInt32("LookupID"),
-                    LookupGroupID = reader.GetInt16("LookupGroupID"),
+                    LookupGroupID = reader.GetInt32("LookupGroupID"),
                     Value = reader.GetString("Value")
                 };
